+++ packages/troika-three-text/src/Typesetter.js
@@ -1,48 +1,97 @@
 /**
+ * @typedef {number|'left'|'center'|'right'} AnchorXValue
+ */
+/**
+ * @typedef {number|'top'|'top-baseline'|'top-cap'|'top-ex'|'middle'|'bottom-baseline'|'bottom'} AnchorYValue
+ */
+
+/**
+ * @typedef {object} TypesetParams
+ * @property {string} text
+ * @property {UserFont|UserFont[]} [font]
+ * @property {string} [lang='en']
+ * @property {number} [sdfGlyphSize=64]
+ * @property {number} [fontSize=1]
+ * @property {number|'normal'|'bold'} [fontWeight='normal']
+ * @property {'normal'|'italic'} [fontStyle='normal']
+ * @property {number} [letterSpacing=0]
+ * @property {'normal'|number} [lineHeight='normal']
+ * @property {number} [maxWidth]
+ * @property {'ltr'|'rtl'} [direction='ltr']
+ * @property {string} [textAlign='left']
+ * @property {number} [textIndent=0]
+ * @property {'normal'|'nowrap'} [whiteSpace='normal']
+ * @property {'normal'|'break-word'} [overflowWrap='normal']
+ * @property {AnchorXValue} [anchorX=0]
+ * @property {AnchorYValue} [anchorY=0]
+ * @property {boolean} [metricsOnly=false]
+ * @property {string} [unicodeFontsURL]
+ * @property {FontResolverResult} [preResolvedFonts]
+ * @property {boolean} [includeCaretPositions=false]
+ * @property {number} [chunkedBoundsSize=8192]
+ * @property {{[rangeStartIndex]: number}} [colorRanges]
+ */
+
+/**
+ * @typedef {object} TypesetResult
+ * @property {Uint16Array} glyphIds id for each glyph, specific to that glyph's font
+ * @property {Uint8Array} glyphFontIndices index into fontData for each glyph
+ * @property {Float32Array} glyphPositions x,y of each glyph's origin in layout
+ * @property {{[font]: {[glyphId]: {path: string, pathBounds: number[]}}}} glyphData data about each glyph appearing in the text
+ * @property {TypesetFontData[]} fontData data about each font used in the text
+ * @property {Float32Array} [caretPositions] startX,endX,bottomY caret positions for each char
+ * @property {Uint8Array} [glyphColors] color for each glyph, if color ranges supplied
+ *         chunkedBounds, //total rects per (n=chunkedBoundsSize) consecutive glyphs
+ *         fontSize, //calculated em height
+ *         topBaseline: anchorYOffset + lines[0].baseline, //y coordinate of the top line's baseline
+ *         blockBounds: [ //bounds for the whole block of text, including vertical padding for lineHeight
+ *           anchorXOffset,
+ *           anchorYOffset - totalHeight,
+ *           anchorXOffset + maxLineWidth,
+ *           anchorYOffset
+ *         ],
+ *         visibleBounds, //total bounds of visible text paths, may be larger or smaller than blockBounds
+ *         timings
+ */
+
+/**
+ * @typedef {object} TypesetFontData
+ * @property src
+ * @property unitsPerEm
+ * @property ascender
+ * @property descender
+ * @property lineHeight
+ * @property capHeight
+ * @property xHeight
+ */
+
+/**
+ * @typedef {function} TypesetterTypesetFunction - compute fonts and layout for some text.
+ * @param {TypesetParams} params
+ * @param {(TypesetResult) => void} callback - function called when typesetting is complete.
+ *    If the params included `preResolvedFonts`, this will be called synchronously.
+ */
+
+/**
+ * @typedef {function} TypesetterMeasureFunction - compute width/height for some text.
+ * @param {TypesetParams} params
+ * @param {(width:number, height:number) => void} callback - function called when measurement is complete.
+ *    If the params included `preResolvedFonts`, this will be called synchronously.
+ */
+
+
+/**
  * Factory function that creates a self-contained environment for processing text typesetting requests.
  *
  * It is important that this function has no closure dependencies, so that it can be easily injected
  * into the source for a Worker without requiring a build step or complex dependency loading. All its
  * dependencies must be passed in at initialization.
  *
- * @param {function} fontParser - a function that accepts an ArrayBuffer of the font data and returns
- * a standardized structure giving access to the font and its glyphs:
- *   {
- *     unitsPerEm: number,
- *     ascender: number,
- *     descender: number,
- *     capHeight: number,
- *     xHeight: number,
- *     lineGap: number,
- *     forEachGlyph(string, fontSize, letterSpacing, callback) {
- *       //invokes callback for each glyph to render, passing it an object:
- *       callback({
- *         index: number,
- *         advanceWidth: number,
- *         xMin: number,
- *         yMin: number,
- *         xMax: number,
- *         yMax: number,
- *         path: string,
- *         pathCommandCount: number
- *       })
- *     }
- *   }
+ * @param {FontResolver} resolveFonts - function to resolve a string to parsed fonts
  * @param {object} bidi - the bidi.js implementation object
- * @param {Object} config
- * @return {Object}
+ * @return {{typeset: TypesetterTypesetFunction, measure: TypesetterMeasureFunction}}
  */
-export function createTypesetter(fontParser, bidi, config) {
-
-  const {
-    defaultFontURL
-  } = config
-
-  /**
-   * Holds parsed font objects by url
-   */
-  const fonts = Object.create(null)
-
+export function createTypesetter(resolveFonts, bidi) {
   const INF = Infinity
 
   // Set of Unicode Default_Ignorable_Code_Point characters, these will not produce visible glyphs
@@ -57,80 +106,49 @@ export function createTypesetter(fontParser, bidi, config) {
   const BREAK_AFTER_CHARS = new RegExp(`${lineBreakingWhiteSpace}|[\\-\\u007C\\u00AD\\u2010\\u2012-\\u2014\\u2027\\u2056\\u2E17\\u2E40]`)
 
   /**
-   * Load a given font url
+   * Load and parse all the necessary fonts to render a given string of text, then group
+   * them into consecutive runs of characters sharing a font.
    */
-  function doLoadFont(url, callback) {
-    function tryLoad() {
-      const onError = err => {
-        console.error(`Failure loading font ${url}${url === defaultFontURL ? '' : '; trying fallback'}`, err)
-        if (url !== defaultFontURL) {
-          url = defaultFontURL
-          tryLoad()
+  function calculateFontRuns({text, lang, fonts, style, weight, preResolvedFonts, unicodeFontsURL}, onDone) {
+    const onResolved = ({chars, fonts: parsedFonts}) => {
+      let curRun, prevVal;
+      const runs = []
+      for (let i = 0; i < chars.length; i++) {
+        if (chars[i] !== prevVal) {
+          prevVal = chars[i];
+          runs.push(curRun = { start: i, end: i, fontObj: parsedFonts[chars[i]]});
+        } else {
+          curRun.end = i;
         }
       }
-      try {
-        const request = new XMLHttpRequest()
-        request.open('get', url, true)
-        request.responseType = 'arraybuffer'
-        request.onload = function () {
-          if (request.status >= 400) {
-            onError(new Error(request.statusText))
-          }
-          else if (request.status > 0) {
-            try {
-              const fontObj = fontParser(request.response)
-              callback(fontObj)
-            } catch (e) {
-              onError(e)
-            }
-          }
-        }
-        request.onerror = onError
-        request.send()
-      } catch(err) {
-        onError(err)
-      }
+      onDone(runs);
     }
-    tryLoad()
-  }
-
-
-  /**
-   * Load a given font url if needed, invoking a callback when it's loaded. If already
-   * loaded, the callback will be called synchronously.
-   */
-  function loadFont(fontUrl, callback) {
-    if (!fontUrl) fontUrl = defaultFontURL
-    let font = fonts[fontUrl]
-    if (font) {
-      // if currently loading font, add to callbacks, otherwise execute immediately
-      if (font.pending) {
-        font.pending.push(callback)
-      } else {
-        callback(font)
-      }
+    if (preResolvedFonts) {
+      onResolved(preResolvedFonts)
     } else {
-      fonts[fontUrl] = {pending: [callback]}
-      doLoadFont(fontUrl, fontObj => {
-        let callbacks = fonts[fontUrl].pending
-        fonts[fontUrl] = fontObj
-        callbacks.forEach(cb => cb(fontObj))
-      })
+      resolveFonts(
+        text,
+        onResolved,
+        { lang, fonts, style, weight, unicodeFontsURL }
+      )
     }
   }
-
 
   /**
    * Main entry point.
    * Process a text string with given font and formatting parameters, and return all info
    * necessary to render all its glyphs.
+   * @type TypesetterTypesetFunction
    */
   function typeset(
     {
       text='',
-      font=defaultFontURL,
+      font,
+      lang='en',
       sdfGlyphSize=64,
-      fontSize=1,
+      fontSize=400,
+      fontWeight=1,
+      fontStyle='normal',
       letterSpacing=0,
       lineHeight='normal',
       maxWidth=INF,
@@ -141,12 +159,14 @@ export function createTypesetter(fontParser, bidi, config) {
       overflowWrap='normal',
       anchorX = 0,
       anchorY = 0,
+      metricsOnly=false,
+      unicodeFontsURL,
+      preResolvedFonts=null,
       includeCaretPositions=false,
       chunkedBoundsSize=8192,
       colorRanges=null
     },
-    callback,
-    metricsOnly=false
+    callback
   ) {
     const mainStart = now()
     const timings = {fontLoad: 0, typesetting: 0}
@@ -164,9 +184,19 @@ export function createTypesetter(fontParser, bidi, config) {
     lineHeight = lineHeight || 'normal'
     textIndent = +textIndent
 
-    loadFont(font, fontObj => {
+    calculateFontRuns({
+      text,
+      lang,
+      style: fontStyle,
+      weight: fontWeight,
+      fonts: typeof font === 'string' ? [{src: font}] : font,
+      unicodeFontsURL,
+      preResolvedFonts
+    }, runs => {
+      timings.fontLoad = now() - mainStart
       const hasMaxWidth = isFinite(maxWidth)
       let glyphIds = null
+      let glyphFontIndices = null
       let glyphPositions = null
       let glyphData = null
       let glyphColors = null
@@ -176,109 +206,160 @@ export function createTypesetter(fontParser, bidi, config) {
       let maxLineWidth = 0
       let renderableGlyphCount = 0
       let canWrap = whiteSpace !== 'nowrap'
-      const {ascender, descender, unitsPerEm, lineGap, capHeight, xHeight} = fontObj
-      timings.fontLoad = now() - mainStart
+      const metricsByFont = new Map() // fontObj -> metrics
       const typesetStart = now()
-
-      // Find conversion between native font units and fontSize units; this will already be done
-      // for the gx/gy values below but everything else we'll need to convert
-      const fontSizeMult = fontSize / unitsPerEm
-
-      // Determine appropriate value for 'normal' line height based on the font's actual metrics
-      // TODO this does not guarantee individual glyphs won't exceed the line height, e.g. Roboto; should we use yMin/Max instead?
-      if (lineHeight === 'normal') {
-        lineHeight = (ascender - descender + lineGap) / unitsPerEm
-      }
-
-      // Determine line height and leading adjustments
-      lineHeight = lineHeight * fontSize
-      const halfLeading = (lineHeight - (ascender - descender) * fontSizeMult) / 2
-      const topBaseline = -(ascender * fontSizeMult + halfLeading)
-      const caretHeight = Math.min(lineHeight, (ascender - descender) * fontSizeMult)
-      const caretBottomOffset = (ascender + descender) / 2 * fontSizeMult - caretHeight / 2
 
       // Distribute glyphs into lines based on wrapping
       let lineXOffset = textIndent
+      let prevRunEndX = 0
       let currentLine = new TextLine()
       const lines = [currentLine]
+      runs.forEach(run => {
+        const { fontObj } = run
+        const { ascender, descender, unitsPerEm, lineGap, capHeight, xHeight } = fontObj
 
-      fontObj.forEachGlyph(text, fontSize, letterSpacing, (glyphObj, glyphX, charIndex) => {
-        const char = text.charAt(charIndex)
-        const glyphWidth = glyphObj.advanceWidth * fontSizeMult
-        const curLineCount = currentLine.count
-        let nextLine
+        // Calculate metrics for each font used
+        let fontData = metricsByFont.get(fontObj)
+        if (!fontData) {
+          // Find conversion between native font units and fontSize units
+          const fontSizeMult = fontSize / unitsPerEm
 
-        // Calc isWhitespace and isEmpty once per glyphObj
-        if (!('isEmpty' in glyphObj)) {
-          glyphObj.isWhitespace = !!char && new RegExp(lineBreakingWhiteSpace).test(char)
-          glyphObj.canBreakAfter = !!char && BREAK_AFTER_CHARS.test(char)
-          glyphObj.isEmpty = glyphObj.xMin === glyphObj.xMax || glyphObj.yMin === glyphObj.yMax || DEFAULT_IGNORABLE_CHARS.test(char)
+          // Determine appropriate value for 'normal' line height based on the font's actual metrics
+          // This does not guarantee individual glyphs won't exceed the line height, e.g. Roboto; should we use yMin/Max instead?
+          const calcLineHeight = lineHeight === 'normal' ?
+            (ascender - descender + lineGap) * fontSizeMult : lineHeight * fontSize
+
+          // Determine line height and leading adjustments
+          const halfLeading = (calcLineHeight - (ascender - descender) * fontSizeMult) / 2
+          const caretHeight = Math.min(calcLineHeight, (ascender - descender) * fontSizeMult)
+          const caretTop = (ascender + descender) / 2 * fontSizeMult + caretHeight / 2
+          fontData = {
+            index: metricsByFont.size,
+            src: fontObj.src,
+            fontObj,
+            fontSizeMult,
+            unitsPerEm,
+            ascender: ascender * fontSizeMult,
+            descender: descender * fontSizeMult,
+            capHeight: capHeight * fontSizeMult,
+            xHeight: xHeight * fontSizeMult,
+            lineHeight: calcLineHeight,
+            baseline: -halfLeading - ascender * fontSizeMult, // baseline offset from top of line height
+            // cap: -halfLeading - capHeight * fontSizeMult, // cap from top of line height
+            // ex: -halfLeading - xHeight * fontSizeMult, // ex from top of line height
+            caretTop: (ascender + descender) / 2 * fontSizeMult + caretHeight / 2,
+            caretBottom: caretTop - caretHeight
+          }
+          metricsByFont.set(fontObj, fontData)
         }
-        if (!glyphObj.isWhitespace && !glyphObj.isEmpty) {
-          renderableGlyphCount++
-        }
+        const { fontSizeMult } = fontData
 
-        // If a non-whitespace character overflows the max width, we need to soft-wrap
-        if (canWrap && hasMaxWidth && !glyphObj.isWhitespace && glyphX + glyphWidth + lineXOffset > maxWidth && curLineCount) {
-          // If it's the first char after a whitespace, start a new line
-          if (currentLine.glyphAt(curLineCount - 1).glyphObj.canBreakAfter) {
-            nextLine = new TextLine()
-            lineXOffset = -glyphX
-          } else {
-            // Back up looking for a whitespace character to wrap at
-            for (let i = curLineCount; i--;) {
-              // If we got the start of the line there's no soft break point; make hard break if overflowWrap='break-word'
-              if (i === 0 && overflowWrap === 'break-word') {
-                nextLine = new TextLine()
-                lineXOffset = -glyphX
-                break
-              }
-              // Found a soft break point; move all chars since it to a new line
-              else if (currentLine.glyphAt(i).glyphObj.canBreakAfter) {
-                nextLine = currentLine.splitAt(i + 1)
-                const adjustX = nextLine.glyphAt(0).x
-                lineXOffset -= adjustX
-                for (let j = nextLine.count; j--;) {
-                  nextLine.glyphAt(j).x -= adjustX
+        const runText = text.slice(run.start, run.end + 1)
+        let prevGlyphX, prevGlyphObj
+        fontObj.forEachGlyph(runText, fontSize, letterSpacing, (glyphObj, glyphX, charIndex) => {
+          glyphX += prevRunEndX
+          charIndex += run.start
+          prevGlyphX = glyphX
+          prevGlyphObj = glyphObj
+          const char = text.charAt(charIndex)
+          const glyphWidth = glyphObj.advanceWidth * fontSizeMult
+          const curLineCount = currentLine.count
+          let nextLine
+
+          // Calc isWhitespace and isEmpty once per glyphObj
+          if (!('isEmpty' in glyphObj)) {
+            glyphObj.isWhitespace = !!char && new RegExp(lineBreakingWhiteSpace).test(char)
+            glyphObj.canBreakAfter = !!char && BREAK_AFTER_CHARS.test(char)
+            glyphObj.isEmpty = glyphObj.xMin === glyphObj.xMax || glyphObj.yMin === glyphObj.yMax || DEFAULT_IGNORABLE_CHARS.test(char)
+          }
+          if (!glyphObj.isWhitespace && !glyphObj.isEmpty) {
+            renderableGlyphCount++
+          }
+
+          // If a non-whitespace character overflows the max width, we need to soft-wrap
+          if (canWrap && hasMaxWidth && !glyphObj.isWhitespace && glyphX + glyphWidth + lineXOffset > maxWidth && curLineCount) {
+            // If it's the first char after a whitespace, start a new line
+            if (currentLine.glyphAt(curLineCount - 1).glyphObj.canBreakAfter) {
+              nextLine = new TextLine()
+              lineXOffset = -glyphX
+            } else {
+              // Back up looking for a whitespace character to wrap at
+              for (let i = curLineCount; i--;) {
+                // If we got the start of the line there's no soft break point; make hard break if overflowWrap='break-word'
+                if (i === 0 && overflowWrap === 'break-word') {
+                  nextLine = new TextLine()
+                  lineXOffset = -glyphX
+                  break
                 }
-                break
+                // Found a soft break point; move all chars since it to a new line
+                else if (currentLine.glyphAt(i).glyphObj.canBreakAfter) {
+                  nextLine = currentLine.splitAt(i + 1)
+                  const adjustX = nextLine.glyphAt(0).x
+                  lineXOffset -= adjustX
+                  for (let j = nextLine.count; j--;) {
+                    nextLine.glyphAt(j).x -= adjustX
+                  }
+                  break
+                }
               }
             }
+            if (nextLine) {
+              currentLine.isSoftWrapped = true
+              currentLine = nextLine
+              lines.push(currentLine)
+              maxLineWidth = maxWidth //after soft wrapping use maxWidth as calculated width
+            }
           }
-          if (nextLine) {
-            currentLine.isSoftWrapped = true
-            currentLine = nextLine
+
+          let fly = currentLine.glyphAt(currentLine.count)
+          fly.glyphObj = glyphObj
+          fly.x = glyphX + lineXOffset
+          fly.width = glyphWidth
+          fly.charIndex = charIndex
+          fly.fontData = fontData
+
+          // Handle hard line breaks
+          if (char === '\n') {
+            currentLine = new TextLine()
             lines.push(currentLine)
-            maxLineWidth = maxWidth //after soft wrapping use maxWidth as calculated width
+            lineXOffset = -(glyphX + glyphWidth + (letterSpacing * fontSize)) + textIndent
           }
-        }
-
-        let fly = currentLine.glyphAt(currentLine.count)
-        fly.glyphObj = glyphObj
-        fly.x = glyphX + lineXOffset
-        fly.width = glyphWidth
-        fly.charIndex = charIndex
-
-        // Handle hard line breaks
-        if (char === '\n') {
-          currentLine = new TextLine()
-          lines.push(currentLine)
-          lineXOffset = -(glyphX + glyphWidth + (letterSpacing * fontSize)) + textIndent
-        }
+        })
+        // At the end of a run we must capture the x position as the starting point for the next run
+        prevRunEndX = prevGlyphX + prevGlyphObj.advanceWidth * fontSizeMult + letterSpacing * fontSize
       })
 
-      // Calculate width of each line (excluding trailing whitespace) and maximum block width
+      // Calculate width/height/baseline of each line (excluding trailing whitespace) and maximum block width
+      let totalHeight = 0
       lines.forEach(line => {
+        let isTrailingWhitespace = true;
         for (let i = line.count; i--;) {
-          let {glyphObj, x, width} = line.glyphAt(i)
-          if (!glyphObj.isWhitespace) {
-            line.width = x + width
+          const glyphInfo = line.glyphAt(i)
+          // omit trailing whitespace from width calculation
+          if (isTrailingWhitespace && !glyphInfo.glyphObj.isWhitespace) {
+            line.width = glyphInfo.x + glyphInfo.width
             if (line.width > maxLineWidth) {
               maxLineWidth = line.width
             }
-            return
+            isTrailingWhitespace = false
           }
+          // use the tallest line height, lowest baseline, and highest cap/ex
+          let {lineHeight, capHeight, xHeight, baseline} = glyphInfo.fontData
+          if (lineHeight > line.lineHeight) line.lineHeight = lineHeight
+          const baselineDiff = baseline - line.baseline
+          if (baselineDiff < 0) { //shift all metrics down
+            line.baseline += baselineDiff
+            line.cap += baselineDiff
+            line.ex += baselineDiff
+          }
+          // compare cap/ex based on new lowest baseline
+          line.cap = Math.max(line.cap, line.baseline + capHeight)
+          line.ex = Math.max(line.ex, line.baseline + xHeight)
         }
+        line.baseline -= totalHeight
+        line.cap -= totalHeight
+        line.ex -= totalHeight
+        totalHeight += line.lineHeight
       })
 
       // Find overall position adjustments for anchoring
@@ -302,15 +383,14 @@ export function createTypesetter(fontParser, bidi, config) {
           anchorYOffset = -anchorY
         }
         else if (typeof anchorY === 'string') {
-          let height = lines.length * lineHeight
           anchorYOffset = anchorY === 'top' ? 0 :
-            anchorY === 'top-baseline' ? -topBaseline :
-            anchorY === 'top-cap' ? -topBaseline - capHeight * fontSizeMult :
-            anchorY === 'top-ex' ? -topBaseline - xHeight * fontSizeMult :
-            anchorY === 'middle' ? height / 2 :
-            anchorY === 'bottom' ? height :
-            anchorY === 'bottom-baseline' ? height - halfLeading + descender * fontSizeMult :
-            parsePercent(anchorY) * height
+            anchorY === 'top-baseline' ? -lines[0].baseline :
+            anchorY === 'top-cap' ? -lines[0].cap :
+            anchorY === 'top-ex' ? -lines[0].ex :
+            anchorY === 'middle' ? totalHeight / 2 :
+            anchorY === 'bottom' ? totalHeight :
+            anchorY === 'bottom-baseline' ? lines[lines.length - 1].baseline :
+            parsePercent(anchorY) * totalHeight
         }
       }
 
@@ -321,13 +401,13 @@ export function createTypesetter(fontParser, bidi, config) {
         // Process each line, applying alignment offsets, adding each glyph to the atlas, and
         // collecting all renderable glyphs into a single collection.
         glyphIds = new Uint16Array(renderableGlyphCount)
+        glyphFontIndices = new Uint8Array(renderableGlyphCount)
         glyphPositions = new Float32Array(renderableGlyphCount * 2)
         glyphData = {}
         visibleBounds = [INF, INF, -INF, -INF]
         chunkedBounds = []
-        let lineYOffset = topBaseline
         if (includeCaretPositions) {
-          caretPositions = new Float32Array(text.length * 3)
+          caretPositions = new Float32Array(text.length * 4)
         }
         if (colorRanges) {
           glyphColors = new Uint8Array(renderableGlyphCount * 3)
@@ -413,7 +493,7 @@ export function createTypesetter(fontParser, bidi, config) {
             let glyphObj
             const setGlyphObj = g => glyphObj = g
             for (let i = 0; i < lineGlyphCount; i++) {
-              let glyphInfo = line.glyphAt(i)
+              const glyphInfo = line.glyphAt(i)
               glyphObj = glyphInfo.glyphObj
               const glyphId = glyphObj.index
 
@@ -422,18 +502,19 @@ export function createTypesetter(fontParser, bidi, config) {
               if (rtl) {
                 const mirrored = bidi.getMirroredCharacter(text[glyphInfo.charIndex])
                 if (mirrored) {
-                  fontObj.forEachGlyph(mirrored, 0, 0, setGlyphObj)
+                  glyphInfo.fontData.fontObj.forEachGlyph(mirrored, 0, 0, setGlyphObj)
                 }
               }
 
               // Add caret positions
               if (includeCaretPositions) {
-                const {charIndex} = glyphInfo
+                const {charIndex, fontData} = glyphInfo
                 const caretLeft = glyphInfo.x + anchorXOffset
                 const caretRight = glyphInfo.x + glyphInfo.width + anchorXOffset
-                caretPositions[charIndex * 3] = rtl ? caretRight : caretLeft //start edge x
-                caretPositions[charIndex * 3 + 1] = rtl ? caretLeft : caretRight //end edge x
-                caretPositions[charIndex * 3 + 2] = lineYOffset + caretBottomOffset + anchorYOffset //common bottom y
+                caretPositions[charIndex * 4] = rtl ? caretRight : caretLeft //start edge x
+                caretPositions[charIndex * 4 + 1] = rtl ? caretLeft : caretRight //end edge x
+                caretPositions[charIndex * 4 + 2] = line.baseline + fontData.caretBottom + anchorYOffset //common bottom y
+                caretPositions[charIndex * 4 + 3] = line.baseline + fontData.caretTop + anchorYOffset //common top y
 
                 // If we skipped any chars from the previous glyph (due to ligature subs), fill in caret
                 // positions for those missing char indices; currently this uses a best-guess by dividing
@@ -460,10 +541,12 @@ export function createTypesetter(fontParser, bidi, config) {
               // Get atlas data for renderable glyphs
               if (!glyphObj.isWhitespace && !glyphObj.isEmpty) {
                 const idx = renderableGlyphIndex++
+                const {fontSizeMult, src: fontSrc, index: fontIndex} = glyphInfo.fontData
 
                 // Add this glyph's path data
-                if (!glyphData[glyphId]) {
-                  glyphData[glyphId] = {
+                const fontGlyphData = glyphData[fontSrc] || (glyphData[fontSrc] = {})
+                if (!fontGlyphData[glyphId]) {
+                  fontGlyphData[glyphId] = {
                     path: glyphObj.path,
                     pathBounds: [glyphObj.xMin, glyphObj.yMin, glyphObj.xMax, glyphObj.yMax]
                   }
@@ -471,7 +554,7 @@ export function createTypesetter(fontParser, bidi, config) {
 
                 // Determine final glyph position and add to glyphPositions array
                 const glyphX = glyphInfo.x + anchorXOffset
-                const glyphY = lineYOffset + anchorYOffset
+                const glyphY = line.baseline + anchorYOffset
                 glyphPositions[idx * 2] = glyphX
                 glyphPositions[idx * 2 + 1] = glyphY
 
@@ -497,8 +580,9 @@ export function createTypesetter(fontParser, bidi, config) {
                 if (visX1 > chunkRect[2]) chunkRect[2] = visX1
                 if (visY1 > chunkRect[3]) chunkRect[3] = visY1
 
-                // Add to glyph ids array
+                // Add to glyph ids and font indices arrays
                 glyphIds[idx] = glyphId
+                glyphFontIndices[idx] = fontIndex
 
                 // Add colors
                 if (colorRanges) {
@@ -510,9 +594,6 @@ export function createTypesetter(fontParser, bidi, config) {
               }
             }
           }
-
-          // Increment y offset for next line
-          lineYOffset -= lineHeight
         })
 
         // Fill in remaining caret positions in case the final character was a ligature
@@ -524,28 +605,30 @@ export function createTypesetter(fontParser, bidi, config) {
         }
       }
 
+      // Assemble final data about each font used
+      const fontData = []
+      metricsByFont.forEach(({index, src, unitsPerEm, ascender, descender, lineHeight, capHeight, xHeight}) => {
+        fontData[index] = {src, unitsPerEm, ascender, descender, lineHeight, capHeight, xHeight}
+      })
+
       // Timing stats
       timings.typesetting = now() - typesetStart
 
       callback({
-        glyphIds, //font indices for each glyph
+        glyphIds, //id for each glyph, specific to that glyph's font
+        glyphFontIndices, //index into fontData for each glyph
         glyphPositions, //x,y of each glyph's origin in layout
         glyphData, //dict holding data about each glyph appearing in the text
+        fontData, //data about each font used in the text
         caretPositions, //startX,endX,bottomY caret positions for each char
-        caretHeight, //height of cursor from bottom to top
+        // caretHeight, //height of cursor from bottom to top - todo per glyph?
         glyphColors, //color for each glyph, if color ranges supplied
         chunkedBounds, //total rects per (n=chunkedBoundsSize) consecutive glyphs
         fontSize, //calculated em height
-        unitsPerEm, //font units per em
-        ascender: ascender * fontSizeMult, //font ascender
-        descender: descender * fontSizeMult, //font descender
-        capHeight: capHeight * fontSizeMult, //font cap-height
-        xHeight: xHeight * fontSizeMult, //font x-height
-        lineHeight, //computed line height
-        topBaseline, //y coordinate of the top line's baseline
+        topBaseline: anchorYOffset + lines[0].baseline, //y coordinate of the top line's baseline
         blockBounds: [ //bounds for the whole block of text, including vertical padding for lineHeight
           anchorXOffset,
-          anchorYOffset - lines.length * lineHeight,
+          anchorYOffset - totalHeight,
           anchorXOffset + maxLineWidth,
           anchorYOffset
         ],
@@ -563,13 +646,13 @@ export function createTypesetter(fontParser, bidi, config) {
    * @param callback
    */
   function measure(args, callback) {
-    typeset(args, (result) => {
+    typeset({...args, metricsOnly: true}, (result) => {
       const [x0, y0, x1, y1] = result.blockBounds
       callback({
         width: x1 - x0,
         height: y1 - y0
       })
-    }, {metricsOnly: true})
+    })
   }
 
   function parsePercent(str) {
@@ -579,15 +662,17 @@ export function createTypesetter(fontParser, bidi, config) {
   }
 
   function fillLigatureCaretPositions(caretPositions, ligStartIndex, ligCount) {
-    const ligStartX = caretPositions[ligStartIndex * 3]
-    const ligEndX = caretPositions[ligStartIndex * 3 + 1]
-    const ligY = caretPositions[ligStartIndex * 3 + 2]
+    const ligStartX = caretPositions[ligStartIndex * 4]
+    const ligEndX = caretPositions[ligStartIndex * 4 + 1]
+    const ligBottom = caretPositions[ligStartIndex * 4 + 2]
+    const ligTop = caretPositions[ligStartIndex * 4 + 3]
     const guessedAdvanceX = (ligEndX - ligStartX) / ligCount
     for (let i = 0; i < ligCount; i++) {
-      const startIndex = (ligStartIndex + i) * 3
+      const startIndex = (ligStartIndex + i) * 4
       caretPositions[startIndex] = ligStartX + guessedAdvanceX * i
       caretPositions[startIndex + 1] = ligStartX + guessedAdvanceX * (i + 1)
-      caretPositions[startIndex + 2] = ligY
+      caretPositions[startIndex + 2] = ligBottom
+      caretPositions[startIndex + 3] = ligTop
     }
   }
 
@@ -599,9 +684,13 @@ export function createTypesetter(fontParser, bidi, config) {
   function TextLine() {
     this.data = []
   }
-  const textLineProps = ['glyphObj', 'x', 'width', 'charIndex']
+  const textLineProps = ['glyphObj', 'x', 'width', 'charIndex', 'fontData']
   TextLine.prototype = {
     width: 0,
+    lineHeight: 0,
+    baseline: 0,
+    cap: 0,
+    ex: 0,
     isSoftWrapped: false,
     get count() {
       return Math.ceil(this.data.length / textLineProps.length)
@@ -634,7 +723,6 @@ export function createTypesetter(fontParser, bidi, config) {
   return {
     typeset,
     measure,
-    loadFont
   }
 }