+++ packages/troika-three-text/src/index.js
@@ -1,6 +1,7 @@
 // Exports for troika-three-text package:
 
-export { configureTextBuilder, typesetterWorkerModule, preloadFont, dumpSDFTextures } from './TextBuilder.js'
+export { configureTextBuilder, getTextRenderInfo, typesetterWorkerModule, preloadFont, dumpSDFTextures } from './TextBuilder.js'
+export { fontResolverWorkerModule } from './FontResolver.js'
 export { Text } from './Text.js'
 export { GlyphsGeometry } from './GlyphsGeometry.js'
 export { createTextDerivedMaterial } from './TextDerivedMaterial.js'