+++ packages/troika-examples/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.48.0-unicode.4](https://github.com/protectwise/troika/compare/v0.48.0-unicode.3...v0.48.0-unicode.4) (2023-09-07)
+
+
+### Features
+
+* allow configuring unicode-font-resolver data URL ([c6ff8d7](https://github.com/protectwise/troika/commit/c6ff8d7e5d5a4e40c73ffb8f9af4c05ee246bec7))
+
+
+
+
+
+# [0.48.0-unicode.3](https://github.com/protectwise/troika/compare/v0.48.0-unicode.2...v0.48.0-unicode.3) (2023-07-10)
+
+**Note:** Version bump only for package troika-examples
+
+
+
+
+
+# [0.48.0-unicode.2](https://github.com/protectwise/troika/compare/v0.48.0-unicode.1...v0.48.0-unicode.2) (2023-07-10)
+
+**Note:** Version bump only for package troika-examples
+
+
+
+
+
+# [0.48.0-unicode.1](https://github.com/protectwise/troika/compare/v0.48.0-unicode.0...v0.48.0-unicode.1) (2023-06-29)
+
+**Note:** Version bump only for package troika-examples
+
+
+
+
+
+# [0.48.0-unicode.0](https://github.com/protectwise/troika/compare/v0.47.2...v0.48.0-unicode.0) (2023-06-29)
+
+
+### Features
+
+* **troika-three-text:** add Arabic and Chinese texts to example ([053687d](https://github.com/protectwise/troika/commit/053687dc03437b67c259b7f4c3b1898f52545dd4))
+* **troika-three-text:** add Noto Emoji font to fallback list ([5cd8c97](https://github.com/protectwise/troika/commit/5cd8c97c4a09f1f57cf6eda0e3eefa15ef0a55b9))
+* **troika-three-text:** integrate unicode-font-resolver for full unicode fallback font coverage ([4c9beef](https://github.com/protectwise/troika/commit/4c9beef4372946239e771b97a0cd35e90ab41ad2))
+
+
+
+
+
 ## [0.47.2](https://github.com/protectwise/troika/compare/v0.47.1...v0.47.2) (2023-05-15)
 
 **Note:** Version bump only for package troika-examples